Change score type to Float64

+++ leaderboard/redis.go
@@ -24,10 +24,10 @@ func NewRedisRepo(redis *redis.Client, name string, pageSize int) Repository {
 }
 
 // RankMember will store a new member score into the leaderboard
-func (r *redisRepo) RankMember(ctx context.Context, username string, score int64) (*User, error) {
+func (r *redisRepo) RankMember(ctx context.Context, username string, score float64) (*User, error) {
 	var user User
-	if val := r.Redis.ZAdd(ctx, r.Name, &redis.Z{Score: float64(score), Member: username}); val == nil {
-		msg := fmt.Sprintf("Failed to insert member %s with score %d to %s", username, score, r.Name)
+	if val := r.Redis.ZAdd(ctx, r.Name, &redis.Z{Score: score, Member: username}); val == nil {
+		msg := fmt.Sprintf("Failed to insert member %s with score %f to %s", username, score, r.Name)
 		err := errors.New(msg)
 		return &user, err
 	}
